Rename SampleData.test to Utils
It is not a test case, just a utility module.
This change is a precursor to moving some shared test functionalities
to this module.
